# [Merged by Bors] - Optimize `process_attestation` with active balance cache