@michaelsproul 2021-09-03
Testing on a Prater node with `--slots-per-restore-point 32` I measured these times:

* ~550ms to load a historic state with this PR, vs 4.6s without
* 500ms to compute the response for the validator inclusion API, vs 5s+ without